feat: 업로드 내역 삭제 완료 시 업로드 내역 refetch
