+++ src/components/uploadRevenue/UploadHistoryTable/UploadHistoryTable.tsx
@@ -1,5 +1,7 @@
 import { useState } from "react";
 
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+
 import ChipButton from "@/components/common/CommonBtns/ChipButton/ChipButton";
 import EmptyData from "@/components/common/EmptyData/EmptyData";
 import AlertModal from "@/components/common/Modals/AlertModal/AlertModal";
@@ -11,6 +13,7 @@ import TableRowUI from "@/components/common/Table/Composition/TableRowUI";
 import useToast from "@/hooks/useToast";
 import { ITransactionUpload } from "@/types/dto";
 import { MODAL_TYPE } from "@/types/enums/modal.enum";
+import { MEMBER_TYPE } from "@/types/enums/user.enum";
 
 interface FileData {
   fileId: number,
@@ -20,9 +23,23 @@ interface FileData {
 const UploadHistroyTable = (
   { uploadList }: { uploadList?: ITransactionUpload[] },
 ) => {
+  const queryClient = useQueryClient();
   const [isCancelUploadModalOpen, setIsCancelUploadModalOpen] = useState(false);
   const [fileData, setFileData] = useState<FileData>({} as FileData);
   const { showToast } = useToast();
+
+  const mutationTest = useMutation(() => {
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve("업로드 내역이 삭제되었습니다.");
+      }, 2000);
+    });
+  }, {
+    onSuccess: (data) => {
+      queryClient.invalidateQueries([MEMBER_TYPE.ADMIN, "settlement-upload-history"]);
+      showToast(data as string);
+    },
+  });
 
   const handleClickCancelUploadBtn = (fileId: number, fileName: string) => {
     setIsCancelUploadModalOpen(true);
@@ -36,9 +53,9 @@ const UploadHistroyTable = (
     // TODO: 업로드 취소 API 작업
     // eslint-disable-next-line no-console
     console.log(`(${fileId}) ${fileName} 파일 삭제`);
+    mutationTest.mutate();
 
     setIsCancelUploadModalOpen(false);
-    showToast("업로드 내역이 삭제되었습니다.");
   };
 
   if (uploadList?.length === 0) {
